Practice_Typer - HTML:JS:Flask/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
python-decouple==3.4
Flask-Toastr
gunicorn==20.1.0
selenium
pytest